refactor: 예외처리 이동
