+++ src/main/java/devgraft/support/exception/StatusConstant.java
@@ -9,7 +9,6 @@ public enum StatusConstant {
     ALREADY_EXISTS_ERROR("요청 정보가 이미 존재합니다.", HttpStatus.BAD_REQUEST),
     UNRELATED_REQUEST_ERROR("자신과 관계 없는 요청은 처리할 수 없습니다.", HttpStatus.FORBIDDEN),
     UNAUTHENTICATED("인증에 실패하였습니다.", HttpStatus.UNAUTHORIZED),
-    HAS_EXPIRED("만료된 인증정보입니다.", HttpStatus.UNAUTHORIZED),
     FORBIDDEN("접근 권한이 없습니다.", HttpStatus.FORBIDDEN),
     NOT_FOUND("요청한 정보를 찾지 못했습니다.", HttpStatus.NOT_FOUND),
     METHOD_NOT_ALLOWED("제공하지 않는 요청입니다.", HttpStatus.METHOD_NOT_ALLOWED),
