+++ src/main/java/devgraft/support/exception/simple/AlreadyExistsException.java
@@ -1,0 +1,14 @@
+package devgraft.support.exception.simple;
+
+import devgraft.support.exception.AbstractRequestException;
+import devgraft.support.exception.StatusConstant;
+
+public class AlreadyExistsException extends AbstractRequestException {
+    public AlreadyExistsException() {
+        super(StatusConstant.ALREADY_EXISTS_ERROR);
+    }
+
+    public AlreadyExistsException(final String message) {
+        super(message, StatusConstant.ALREADY_EXISTS_ERROR);
+    }
+}
