+++ src/main/java/devgraft/support/exception/simple/UnrelatedException.java
@@ -7,4 +7,8 @@ public class UnrelatedException extends AbstractRequestException {
     public UnrelatedException() {
         super(StatusConstant.UNRELATED_REQUEST_ERROR);
     }
+
+    public UnrelatedException(final String message) {
+        super(message, StatusConstant.UNRELATED_REQUEST_ERROR);
+    }
 }
