+++ src/test/java/devgraft/friend/app/CancelFriendServiceTest.java
@@ -1,5 +1,6 @@
 package devgraft.friend.app;
 
+import devgraft.friend.app.exception.NotFoundFriendRelationException;
 import devgraft.friend.domain.FriendEventSender;
 import devgraft.friend.domain.FriendRelationFixture;
 import devgraft.friend.domain.FriendRelationRepository;
